refactor(rag): clarify ingested document structure api
