+++ app/lib/rag.py
@@ -3,7 +3,6 @@
 All plugins and external modules import RAG functionality from here.
 """
 
-from app.rag import structure
 from app.rag.exceptions import (
     DocumentNotFoundError,
     RAGNotReadyError,
@@ -15,6 +14,7 @@ from app.rag.ingestion import ingest_document
 from app.rag.retrieval import agentic_retrieve_context
 from app.rag.store import copy_document_chunk_links
 from app.rag.types import DocumentSection, RetrievedChunk
+from app.rag.utils import get_rag_ingested_document_structure
 
 __all__ = [
     "DocumentNotFoundError",
@@ -25,12 +25,7 @@ __all__ = [
     "ingest_document",
     "agentic_retrieve_context",
     "copy_document_chunk_links",
-    "get_document_structure",
+    "get_rag_ingested_document_structure",
     "DocumentSection",
     "RetrievedChunk",
 ]
-
-
-async def get_document_structure(user_id: int, document_id: int) -> list[DocumentSection]:
-    """Return ordered RAG section metadata for a document."""
-    return await structure.get_document_structure(user_id, document_id)
